增加 CN_UPDATE=false 环境变量
设置为true会使用国内pip源和国内alpine软件源

+++ docker/entrypoint.sh
@@ -8,8 +8,10 @@ if [ "$NASTOOL_AUTO_UPDATE" = "true" ]; then
     if [ ! -s /tmp/third_party.txt.sha256sum ]; then
         sha256sum third_party.txt > /tmp/third_party.txt.sha256sum
     fi
-    if [ ! -s /tmp/package_list.txt.sha256sum ]; then
-        sha256sum package_list.txt > /tmp/package_list.txt.sha256sum
+    if [ "$NASTOOL_VERSION" = "full" ]; then
+        if [ ! -s /tmp/package_list.txt.sha256sum ]; then
+            sha256sum package_list.txt > /tmp/package_list.txt.sha256sum
+        fi
     fi
     echo "更新程序..."
     git remote set-url origin ${REPO_URL} &>/dev/null
@@ -25,8 +27,13 @@ if [ "$NASTOOL_AUTO_UPDATE" = "true" ]; then
         hash_new=$(sha256sum requirements.txt)
         if [ "$hash_old" != "$hash_new" ]; then
             echo "检测到requirements.txt有变化，重新安装依赖..."
-            pip install --upgrade pip setuptools wheel
-            pip install -r requirements.txt
+            if [ "$CN_UPDATE" = "true" ]; then
+                pip install --upgrade pip setuptools wheel -i https://pypi.tuna.tsinghua.edu.cn/simple
+                pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple
+            else
+                pip install --upgrade pip setuptools wheel
+                pip install -r requirements.txt
+            fi
             if [ $? -ne 0 ]; then
                 echo "无法安装依赖，请更新镜像..."
             else
@@ -47,18 +54,24 @@ if [ "$NASTOOL_AUTO_UPDATE" = "true" ]; then
             fi
         fi
         # 系统软件包更新
-        hash_old=$(cat /tmp/package_list.txt.sha256sum)
-        hash_new=$(sha256sum package_list.txt)
-        if [ "$hash_old" != "$hash_new" ]; then
-            echo "检测到package_list.txt有变化，更新软件包..."
-            apk add --no-cache libffi-dev
-            apk add --no-cache $(echo $(cat package_list.txt))
-            if [ $? -ne 0 ]; then
-                echo "无法更新软件包，请更新镜像..."
-            else
-                apk del libffi-dev
-                echo "软件包安装成功..."
-                sha256sum package_list.txt > /tmp/package_list.txt.sha256sum
+        if [ "$NASTOOL_VERSION" = "full" ]; then
+            hash_old=$(cat /tmp/package_list.txt.sha256sum)
+            hash_new=$(sha256sum package_list.txt)
+            if [ "$hash_old" != "$hash_new" ]; then
+                echo "检测到package_list.txt有变化，更新软件包..."
+                if [ "$CN_UPDATE" = "true" ]; then
+                    sed -i 's/dl-cdn.alpinelinux.org/mirrors.ustc.edu.cn/g' /etc/apk/repositories
+                    apk update -f
+                fi
+                apk add --no-cache libffi-dev
+                apk add --no-cache $(echo $(cat package_list.txt))
+                if [ $? -ne 0 ]; then
+                    echo "无法更新软件包，请更新镜像..."
+                else
+                    apk del libffi-dev
+                    echo "软件包安装成功..."
+                    sha256sum package_list.txt > /tmp/package_list.txt.sha256sum
+                fi
             fi
         fi
     else
@@ -67,17 +80,17 @@ if [ "$NASTOOL_AUTO_UPDATE" = "true" ]; then
 else
     echo "程序自动升级已关闭，如需自动升级请在创建容器时设置环境变量：NASTOOL_AUTO_UPDATE=true"
 fi
+
 echo "以PUID=${PUID}，PGID=${PGID}的身份启动程序..."
-mkdir -p /.local
-mkdir -p /.pm2
+
 if [ "$NASTOOL_VERSION" = "lite" ]; then
-  chown -R "${PUID}":"${PGID}" "${WORKDIR}" /config /.local /.pm2
+    mkdir -p /.pm2
+    chown -R "${PUID}":"${PGID}" "${WORKDIR}" /config /.local /.pm2
 else
-  chown -R "${PUID}":"${PGID}" "${WORKDIR}" /config /usr/lib/chromium /.local /.pm2
-  export PATH=$PATH:/usr/lib/chromium
+    mkdir -p /.local
+    mkdir -p /.pm2
+    chown -R "${PUID}":"${PGID}" "${WORKDIR}" /config /usr/lib/chromium /.local /.pm2
+    export PATH=$PATH:/usr/lib/chromium
 fi
 umask "${UMASK}"
-if ! which dumb-init; then
-    apk add --no-cache dumb-init
-fi
 exec su-exec "${PUID}":"${PGID}" "$(which dumb-init)" "$(which pm2-runtime)" start run.py -n NAStool --interpreter python3
